% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Utilities.R
\name{AggregateByCluster}
\alias{AggregateByCluster}
\title{Aggregate a single-cell expression matrix.}
\usage{
AggregateByCluster(expr, group_labels, method, scale_genes = F)
}
\arguments{
\item{expr}{Single-cell expression matrix with one cell per column.
This function expects untransformed normalized expression, e.g. counts
per million.}

\item{group_labels}{Categories within which to aggregate expression.}

\item{method}{How to aggregate. detection_rate finds the proportion of cells
expressing a given gene in a given group_labels. Otherwise, values from expr are
simply averaged.}

\item{scale_genes}{Should each gene be scaled to a max of 1?}
}
\description{
Aggregate a single-cell expression matrix.
}
