% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Transparency.R
\name{CheckInputs}
\alias{CheckInputs}
\title{Helper function for commonly used inputs.}
\usage{
CheckInputs(
  query,
  reference,
  pairs = NULL,
  compute_penalty = "correlation_distance",
  do_match_ncol
)
}
\description{
Helper function for commonly used inputs.
}
