% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Transparency.R
\name{ClusterGenes}
\alias{ClusterGenes}
\title{Helper function for EvaluateByGene.}
\usage{
ClusterGenes(query, reference, K = 20)
}
\arguments{
\item{query}{@param reference Expression matrices to be matched.
Each column should be one cluster. Each row should be one gene or genomic feature.
The genes should be shared and ordered the same way.}

\item{K}{number of clusters}
}
\description{
Helper function for EvaluateByGene.
}
