% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DiscreteFunctionTools.R
\name{ComputeSingleCoordPaths}
\alias{ComputeSingleCoordPaths}
\title{At each coordinate of init, cycle through all possible values and record the output.}
\usage{
ComputeSingleCoordPaths(FUN, values, init, niter = 10, ...)
}
\arguments{
\item{FUN}{Function to minimize. Should take a list or vector as the first input.}

\item{values}{What to iterate over at each coordinate.}

\item{init}{perturbations of each coordinate of this input are studied.}

\item{...}{Additional args passed to FUN.}
}
\description{
At each coordinate of init, cycle through all possible values and record the output.
}
