% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DiscreteFunctionTools.R
\name{CoordinateDescent}
\alias{CoordinateDescent}
\title{Minimize the given function by coordinate descent on a finite domain.}
\usage{
CoordinateDescent(FUN, values, init = NULL, niter = 10, verbose = T, ...)
}
\arguments{
\item{FUN}{Function to minimize. Should take a list or vector as the first input.}

\item{values}{What to iterate over at each coordinate.}

\item{init}{Initialization -- first list to put into FUN.}

\item{verbose}{If true, print optimization updates.}

\item{...}{Additional args passed to FUN.}
}
\description{
Minimize the given function by coordinate descent on a finite domain.
}
