% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Transparency.R
\name{CountSupportingGenes}
\alias{CountSupportingGenes}
\title{Count genes that support each matching decision.}
\usage{
CountSupportingGenes(gene_eval)
}
\arguments{
\item{gene_eval}{Results of EvaluateByGene, or RunCellMatch(...)$eval_results .}
}
\description{
Count genes that support each matching decision.
}
