% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Transparency.R
\name{EvaluateByGene}
\alias{EvaluateByGene}
\title{Evaluate matches between conformable count matrices.}
\usage{
EvaluateByGene(
  query,
  reference,
  equivalents = colnames(reference),
  compute_penalty = "correlation_distance",
  fold_change_cutoff = 2,
  results_path = NULL,
  do_heatmaps = !is.null(results_path),
  reorder_genes = F,
  pseudocount = 0.1,
  ...
)
}
\arguments{
\item{query}{@param reference Expression matrices to be matched.
Each column should be one cluster. Each row should be one gene or genomic feature.
The genes should be shared and ordered the same way.}

\item{equivalents}{a list of columns from reference to be matched with the query.}

\item{results_path}{where to save heatmaps}

\item{do_heatmaps}{obvious}

\item{...}{Additional parameters passed to ClusterGenes}
}
\description{
Evaluate matches between conformable count matrices.
}
