% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Robustness.R
\name{ExamineRobustness}
\alias{ExamineRobustness}
\title{Check different genesets, distance functions, and gene-wise data transformations.}
\usage{
ExamineRobustness(
  genesets,
  distances_use = GetDefaultDistances(),
  transforms_use = GetDefaultTransformations(),
  verbose = T
)
}
\arguments{
\item{genesets}{Named list of character vectors}

\item{distances_use}{Named list of functions compatible with the
compute_penalty argument of SelectModel. See examples in GetDefaultDistances().}

\item{transforms_use}{Named list of functions. See examples in GetDefaultTransformations().}

\item{verbose}{Logical. If TRUE, print progress.}
}
\description{
Check different genesets, distance functions, and gene-wise data transformations.
}
