% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Transparency.R
\name{FullComparisonReport}
\alias{FullComparisonReport}
\title{Call transparency functions on high-scoring models.}
\usage{
FullComparisonReport(neighboring_model_scores, results_path)
}
\arguments{
\item{neighboring_model_scores}{Output from MutateModel.}

\item{results_path}{Where to store plots.}
}
\description{
Call transparency functions on high-scoring models.
}
