% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Utilities.R
\name{get_cpm}
\alias{get_cpm}
\title{Extract counts per million, not transformed, from a Seurat version 2 object.}
\usage{
get_cpm(seurat_object)
}
\description{
Extract counts per million, not transformed, from a Seurat version 2 object.
}
