% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SpeciesConvert.R
\name{get_ortholog}
\alias{get_ortholog}
\title{Return the ortholog of a given gene or NA (if no match).
Human and mouse only so far.}
\usage{
get_ortholog(x, from, to, ...)
}
\description{
Return the ortholog of a given gene or NA (if no match).
Human and mouse only so far.
}
