% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Robustness.R
\name{GetDefaultDistances}
\alias{GetDefaultDistances}
\title{Return a named list of methods for distance computation.}
\usage{
GetDefaultDistances()
}
\description{
Return a named list of methods for distance computation.
}
