% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Robustness.R
\name{GetDefaultTransformations}
\alias{GetDefaultTransformations}
\title{Return a named list of methods for transforming input.}
\usage{
GetDefaultTransformations()
}
\description{
Return a named list of methods for transforming input.
}
