% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ModelSelection.R
\name{GetReasonableAlternative}
\alias{GetReasonableAlternative}
\title{Return another matching model that seems just as good.}
\usage{
GetReasonableAlternative(neighboring_model_scores)
}
\arguments{
\item{neighboring_model_scores}{Output from MutateModel.}
}
\description{
Return another matching model that seems just as good.
}
