% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Utilities.R
\name{MergeMatrixList}
\alias{MergeMatrixList}
\title{Merge matrices assuming that rownames are overlapping but non-identical sets of genes.
For genes present in only some matrices, zeroes are added in the other matrices.}
\usage{
MergeMatrixList(expr)
}
\arguments{
\item{expr}{List of matrices}
}
\description{
Merge matrices assuming that rownames are overlapping but non-identical sets of genes.
For genes present in only some matrices, zeroes are added in the other matrices.
}
