% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ModelSelection.R
\name{MutateModel}
\alias{MutateModel}
\title{Study all neighboring matching models by "mutating" the model at each stage.}
\usage{
MutateModel(
  query,
  reference,
  do_omit = F,
  compute_penalty = "correlation_distance",
  penalty_is_genewise = T,
  values = colnames(reference),
  init,
  ...
)
}
\arguments{
\item{do_omit}{Boolean. See \code{?StudyModel}. When do_omit is false (default),
StudyModel is called with ComputeSingleCoordPaths as wrap_fun. Otherwise,
StudyModel is called with OmitSingleCoords as wrap_fun.}
}
\description{
Study all neighboring matching models by "mutating" the model at each stage.
}
