% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DiscreteFunctionTools.R
\name{OmitSingleCoords}
\alias{OmitSingleCoords}
\title{At each coordinate of init, omit that coordinate and record the output.}
\usage{
OmitSingleCoords(FUN, init = NULL, values = NULL, ...)
}
\arguments{
\item{FUN}{Function to minimize. Should take a list or vector as the first input.
Must also accept an optional argument 'index_omit' giving the coordinate to omit.
The function itself must know how to omit a coordinate in a manner appropriate to the context.}

\item{init}{Initialization -- first list to put into FUN.}

\item{values}{Ignored. This is only present for consistency with similar functions.}

\item{...}{Additional args passed to FUN.}
}
\description{
At each coordinate of init, omit that coordinate and record the output.
}
