% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Utilities.R
\name{PadMatrix}
\alias{PadMatrix}
\title{Pad a matrix with zeroes to facilitate merging of expression data with
non-identical gene-sets.}
\usage{
PadMatrix(X, desired_rownames)
}
\description{
Pad a matrix with zeroes to facilitate merging of expression data with
non-identical gene-sets.
}
