% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Transparency.R
\name{PairedHeatmap}
\alias{PairedHeatmap}
\title{Depict conformable count matrices.}
\usage{
PairedHeatmap(
  query,
  reference,
  genes,
  reorder_genes = TRUE,
  genes_to_label = ifelse(length(genes) > 50, "none", "all"),
  query_prefix = "q_",
  reference_prefix = "r_",
  include_difference = F,
  scores = NULL
)
}
\arguments{
\item{query}{@param reference Expression matrices to be matched.
Each column should be one cluster. Each row should be one gene or genomic feature.
The dimensions should be equal.}

\item{genes}{Genes to show.}

\item{reorder_genes}{If true, use the seriation TSP method to reorder the genes.}

\item{genes_to_label}{"none", "all", or a character vector of gene names.}

\item{query_prefix}{@param reference_prefix Prefixes to distinguish column labels. Must be different.}

\item{include_difference}{Boolean. Include a third panel with query minus reference?}

\item{scores}{scores. If provided, these are shown as a marginal barplot. The plot is
visually optimized for situations where the scores range from 0 to 1.}
}
\description{
Depict conformable count matrices.
}
