% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Utilities.R
\name{PlotEquivalentsTSNE}
\alias{PlotEquivalentsTSNE}
\title{Show equivalents on a tSNE or similar.}
\usage{
PlotEquivalentsTSNE(
  embedding,
  group_labels,
  pairs,
  background_color = "grey",
  background_name = "background_"
)
}
\arguments{
\item{embedding}{tSNE, UMAP, or embedding of choice. Data frame or matrix with two named columns.}

\item{group_labels}{Character vector of same size and embedding.}

\item{pairs}{Small dataframe with columns query, reference, and (optionally) color.

This function is for when you have a tSNE for a \emph{reference} dataset, with
cluster labels from the \emph{reference} dataset. It will show where the query falls
on this reference visualization. If you want to do the opposite, i.e. start with a
query dataset and use a reference color-scheme, you'll have to trick the function
by swapping the column names in the pairs argument.

Reference cells assigned multiple query equivalents will be colored
as one or another at random.}
}
\description{
Show equivalents on a tSNE or similar.
}
