% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Transparency.R
\name{PlotMissingGenes}
\alias{PlotMissingGenes}
\title{Display (and save to disk) missing genes from a given cell type.}
\usage{
PlotMissingGenes(
  query,
  reference,
  query_baseline = colnames(query)[[1]],
  reference_baseline = colnames(reference)[[1]],
  query_mature = colnames(query)[[ncol(query)]],
  reference_mature = colnames(reference)[[ncol(reference)]],
  results_path,
  genes = rownames(reference),
  genes_label = NULL,
  fc_cutoff = 2,
  cpm_cutoff = 1,
  prop_genes_label = 0.01,
  main = "Change from baseline to mature state"
)
}
\arguments{
\item{query}{@param reference normalized count matrices sharing rownames (gene names).}

\item{query_baseline}{@param reference_baseline Character. Baseline for each species or condition: for example, pluripotent cells.
Defaults to first column of query or reference respectively.}

\item{query_mature}{@param reference_mature Character. Clusters to be compared when looking for missing genes.
Defaults to last column of query or reference respectively.}
}
\description{
Display (and save to disk) missing genes from a given cell type.
}
