% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ModelSelection.R
\name{PlotNeighboringModels}
\alias{PlotNeighboringModels}
\title{Display output from MutateModel.}
\usage{
PlotNeighboringModels(
  neighboring_model_scores,
  query_order = NULL,
  ref_order = NULL,
  omit_one_model_scores = NULL,
  omit_one_display_mode = "both",
  results_path,
  plotname = "alternatives.pdf"
)
}
\arguments{
\item{neighboring_model_scores}{Output from MutateModel (default mode)}

\item{omit_one_model_scores}{Output from MutateModel (with do_omit=T)}

\item{omit_one_display_mode}{How to display scores of omitted models. One of c("sidebar", "main", "both").}

\item{results_path}{Where to save output.}

\item{plotname}{filename of plot. Should end in a ggplot-supported file extension
such as .png or .pdf.}
}
\description{
Display output from MutateModel.
}
