% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ModelSelection.R
\name{RunCellMatch}
\alias{RunCellMatch}
\title{Match query and reference RNA profiles and render a full report.}
\usage{
RunCellMatch(
  query,
  reference,
  results_path,
  K = 2000,
  num_init = 5,
  verbose = T
)
}
\arguments{
\item{query}{@param reference Expression matrices to be matched.
Each column should be one cluster. Each row should be one gene or genomic feature.
The genes should be shared and ordered the same way.}

\item{results_path}{Where to save outputted files.}

\item{K}{How many genes to select.}

\item{num_init}{Number of random restarts for coordinate descent. Higher means slower and better quality.}

\item{compute_penalty}{a gene-wise penalty function acting on two matrices of the same size,
and producing numeric vector output with one entry per gene. Make sure big is bad.
Accepts shorthand: try "correlation_distance" (recommended), "squared", or "absolute".}
}
\description{
Match query and reference RNA profiles and render a full report.
}
