% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ModelSelection.R
\name{SampleModels}
\alias{SampleModels}
\title{Randomly draw \code{niter} models and compute their penalty functions.}
\usage{
SampleModels(
  query,
  reference,
  compute_penalty = "correlation_distance",
  penalty_is_genewise = T,
  values = colnames(reference),
  init = sample(values, size = ncol(query), replace = T),
  niter = 100,
  ...
)
}
\description{
See \code{?StudyModel} and \code{?SampleUniformly}.
}
