% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DiscreteFunctionTools.R
\name{SampleUniformly}
\alias{SampleUniformly}
\title{Sample inputs uniformly at random.}
\usage{
SampleUniformly(FUN, values, init, niter, coords_fix = NULL, seed = 0, ...)
}
\arguments{
\item{FUN}{Function to minimize. Should take a list or vector as the first input.}

\item{values}{What to iterate over at each coordinate.}

\item{init}{If coords_fix is given, \code{init[coords_fix]} remains under the null model and only
the remaining entries are sampled.
Otherwise, only the length matters, as it determines the dimension of the cube to be sampled.}

\item{niter}{How many to sample.}

\item{coords_fix}{Coordinates to leave fixed. This can generate a different null distribution.}

\item{seed}{input to set.seed (for repeatable results).}

\item{...}{Additional args passed to FUN.}
}
\description{
Sample inputs uniformly at random.
}
