% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ModelSelection.R
\name{SelectModelFast}
\alias{SelectModelFast}
\title{Form matches between columns of two count matrices.}
\usage{
SelectModelFast(
  query,
  reference,
  compute_penalty = "correlation_distance",
  penalty_is_genewise = T,
  values = colnames(reference),
  init = sample(values, ncol(query), replace = T),
  ...
)
}
\description{
Use coordinate descent to select a (locally) optimal matching model.
See \code{?StudyModel}; this calls that with CoordinateDescent as wrap_fun.
See also \code{?SelectModelThoroughly}.
}
