% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ModelSelection.R
\name{SelectModelThoroughly}
\alias{SelectModelThoroughly}
\title{Form matches between columns of two count matrices.}
\usage{
SelectModelThoroughly(
  query,
  reference,
  compute_penalty = "correlation_distance",
  penalty_is_genewise = T,
  values = colnames(reference),
  num_init = 5,
  ...
)
}
\description{
Works like SelectModelFast but with multiple random initializations.
}
