% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ModelSelection.R
\name{StudyModel}
\alias{StudyModel}
\title{Tools for model selection and comparison, implemented via DiscreteFunctionTools.}
\usage{
StudyModel(
  query,
  reference,
  compute_penalty = "correlation_distance",
  penalty_is_genewise = T,
  wrap_fun,
  values = colnames(reference),
  init = sample(values, ncol(query), replace = T),
  ...
)
}
\arguments{
\item{query}{@param reference Expression matrices to be matched.
Each column should be one cluster. Each row should be one gene or genomic feature.
The genes should be shared and ordered the same way.}

\item{compute_penalty}{a gene-wise penalty function acting on two matrices of the same size,
and producing numeric vector output with one entry per gene. Make sure big is bad.
Accepts shorthand: try "correlation_distance" (recommended), "squared", or "absolute".}

\item{penalty_is_genewise}{For complete flexibility with the penalty, set this to FALSE and
pass in any penalty function that returns a scalar. Make sure big is bad.}

\item{wrap_fun}{Routine to wrap around the outside.
Try CoordinateDescent, ComputeSingleCoordPaths, or SampleUniformly.}

\item{values}{@param init @param ... Passed to wrap_fun.}
}
\description{
Tools for model selection and comparison, implemented via DiscreteFunctionTools.
}
